+++ pkgs/cli_config/test/cli_config_test.dart
@@ -371,11 +371,20 @@ void main() {
     expect(() => Config.fromConfigFileContents(fileContents: "['asdf']"),
         throwsFormatException);
     expect(
-      () => Config.fromConfigFileContents(fileContents: '''foo:
-  bar:
-    WRONGKEY:
-      1
-'''),
+      () => Config.fromConfigFileContents(
+          fileContents: '''
+WRONGKEY:
+  1
+'''
+              .trim()),
+      throwsFormatException,
+    );
+    expect(
+      () => Config.fromConfigFileContents(
+          fileContents: '''
+1: 'asdf'
+'''
+              .trim()),
       throwsFormatException,
     );
   });
@@ -503,5 +512,47 @@ void main() {
       expect(config.optionalPathList('my_list'), <String>[]);
       expect(config.pathList('my_list'), <String>[]);
     }
+  });
+
+  test('non-string key maps', () {
+    // This is valid in YAML (not in JSON).
+    //
+    // Such values cannot be accessed with our hierarchical keys, but they can
+    // be accessed with [Config.valueOf].
+    final config = Config(
+      fileParsed: {
+        'my_non_string_key_map': {
+          1: 'asdf',
+          2: 'foo',
+        },
+      },
+    );
+
+    expect(
+      config.valueOf<Map<Object, Object?>>('my_non_string_key_map'),
+      {
+        1: 'asdf',
+        2: 'foo',
+      },
+    );
+  });
+
+  test('null values in maps', () {
+    final config = Config(
+      fileParsed: {
+        'my_non_string_key_map': {
+          'x': null,
+          'y': 42,
+        },
+      },
+    );
+
+    expect(
+      config.valueOf<Map<Object, Object?>>('my_non_string_key_map'),
+      {
+        'x': null,
+        'y': 42,
+      },
+    );
   });
 }